requirements.txt
fastapi==0.96.0
firebase_admin==6.1.0
numpy==1.23.5
protobuf==4.23.2
tensorflow==2.12.0
uvicorn==0.22.0
google.cloud